src/main/webapp/app/entities/report-intervention-list/delete/report-intervention-list-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { ReportInterventionListService } from '../service/report-intervention-list.service';

import { ReportInterventionListDeleteDialogComponent } from './report-intervention-list-delete-dialog.component';

describe('ReportInterventionList Management Delete Component', () => {
  let comp: ReportInterventionListDeleteDialogComponent;
  let fixture: ComponentFixture<ReportInterventionListDeleteDialogComponent>;
  let service: ReportInterventionListService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [ReportInterventionListDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(ReportInterventionListDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(ReportInterventionListDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(ReportInterventionListService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
